ControlClient: Move exceptions to a detected file.
